+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/CodyCommandMode.kt
@@ -1,8 +1,5 @@
 @file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
 package com.sourcegraph.cody.protocol_generated;
 
-data class SearchPanelFile(
-  val uri: Uri,
-  val snippets: List<SearchPanelSnippet>,
-)
+typealias CodyCommandMode = String // One of: ask, edit, insert