+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/WebviewMessage.kt
@@ -22,6 +22,7 @@ sealed class WebviewMessage {
           "restoreHistory" -> context.deserialize<RestoreHistoryWebviewMessage>(element, RestoreHistoryWebviewMessage::class.java)
           "deleteHistory" -> context.deserialize<DeleteHistoryWebviewMessage>(element, DeleteHistoryWebviewMessage::class.java)
           "links" -> context.deserialize<LinksWebviewMessage>(element, LinksWebviewMessage::class.java)
+          "openURI" -> context.deserialize<OpenURIWebviewMessage>(element, OpenURIWebviewMessage::class.java)
           "show-page" -> context.deserialize<`show-pageWebviewMessage`>(element, `show-pageWebviewMessage`::class.java)
           "chatModel" -> context.deserialize<ChatModelWebviewMessage>(element, ChatModelWebviewMessage::class.java)
           "get-chat-models" -> context.deserialize<`get-chat-modelsWebviewMessage`>(element, `get-chat-modelsWebviewMessage`::class.java)
@@ -41,8 +42,6 @@ sealed class WebviewMessage {
           "simplified-onboarding" -> context.deserialize<`simplified-onboardingWebviewMessage`>(element, `simplified-onboardingWebviewMessage`::class.java)
           "getUserContext" -> context.deserialize<GetUserContextWebviewMessage>(element, GetUserContextWebviewMessage::class.java)
           "queryContextItems" -> context.deserialize<QueryContextItemsWebviewMessage>(element, QueryContextItemsWebviewMessage::class.java)
-          "search" -> context.deserialize<SearchWebviewMessage>(element, SearchWebviewMessage::class.java)
-          "show-search-result" -> context.deserialize<`show-search-resultWebviewMessage`>(element, `show-search-resultWebviewMessage`::class.java)
           "reset" -> context.deserialize<ResetWebviewMessage>(element, ResetWebviewMessage::class.java)
           "attribution-search" -> context.deserialize<`attribution-searchWebviewMessage`>(element, `attribution-searchWebviewMessage`::class.java)
           "troubleshoot/reloadAuth" -> context.deserialize<Troubleshoot_reloadAuthWebviewMessage>(element, Troubleshoot_reloadAuthWebviewMessage::class.java)
@@ -151,6 +150,16 @@ data class LinksWebviewMessage(
 
   enum class CommandEnum {
     @SerializedName("links") Links,
+  }
+}
+
+data class OpenURIWebviewMessage(
+  val command: CommandEnum, // Oneof: openURI
+  val uri: Uri,
+) : WebviewMessage() {
+
+  enum class CommandEnum {
+    @SerializedName("openURI") OpenURI,
   }
 }
 
@@ -304,7 +313,7 @@ data class CopyWebviewMessage(
 
 data class AuthWebviewMessage(
   val command: CommandEnum, // Oneof: auth
-  val authKind: AuthKindEnum, // Oneof: signin, signout, support, callback, simplified-onboarding
+  val authKind: AuthKindEnum, // Oneof: signin, signout, support, callback, simplified-onboarding, offline
   val endpoint: String? = null,
   val value: String? = null,
   val authMethod: AuthMethod? = null, // Oneof: dotcom, github, gitlab, google
@@ -320,6 +329,7 @@ data class AuthWebviewMessage(
     @SerializedName("support") Support,
     @SerializedName("callback") Callback,
     @SerializedName("simplified-onboarding") `Simplified-onboarding`,
+    @SerializedName("offline") Offline,
   }
 }
 
@@ -363,27 +373,6 @@ data class QueryContextItemsWebviewMessage(
 
   enum class CommandEnum {
     @SerializedName("queryContextItems") QueryContextItems,
-  }
-}
-
-data class SearchWebviewMessage(
-  val command: CommandEnum, // Oneof: search
-  val query: String,
-) : WebviewMessage() {
-
-  enum class CommandEnum {
-    @SerializedName("search") Search,
-  }
-}
-
-data class `show-search-resultWebviewMessage`(
-  val command: CommandEnum, // Oneof: show-search-result
-  val uri: Uri,
-  val range: RangeData,
-) : WebviewMessage() {
-
-  enum class CommandEnum {
-    @SerializedName("show-search-result") `Show-search-result`,
   }
 }