+++ lib/shared/src/auth/types.ts
@@ -9,6 +9,11 @@ export interface AuthStatus {
     endpoint: string | null
     isDotCom: boolean
     isLoggedIn: boolean
+    /**
+     * Used to enable Fireworks tracing for Sourcegraph teammates on DotCom.
+     * https://readme.fireworks.ai/docs/enabling-tracing
+     */
+    isFireworksTracingEnabled: boolean
     showInvalidAccessTokenError: boolean
     authenticated: boolean
     hasVerifiedEmail: boolean
@@ -29,6 +34,8 @@ export interface AuthStatus {
      * buttons in the UI.
      */
     userCanUpgrade: boolean
+
+    isOfflineMode?: boolean
 }
 
 export interface AuthStatusProvider {
@@ -39,6 +46,7 @@ export const defaultAuthStatus = {
     endpoint: '',
     isDotCom: true,
     isLoggedIn: false,
+    isFireworksTracingEnabled: false,
     showInvalidAccessTokenError: false,
     authenticated: false,
     hasVerifiedEmail: false,
@@ -57,6 +65,7 @@ export const unauthenticatedStatus = {
     endpoint: '',
     isDotCom: true,
     isLoggedIn: false,
+    isFireworksTracingEnabled: false,
     showInvalidAccessTokenError: true,
     authenticated: false,
     hasVerifiedEmail: false,
@@ -76,6 +85,7 @@ export const networkErrorAuthStatus = {
     showInvalidAccessTokenError: false,
     authenticated: false,
     isLoggedIn: false,
+    isFireworksTracingEnabled: false,
     hasVerifiedEmail: false,
     showNetworkError: true,
     requiresVerifiedEmail: false,
@@ -88,3 +98,23 @@ export const networkErrorAuthStatus = {
     avatarURL: '',
     codyApiVersion: 0,
 } satisfies Omit<AuthStatus, 'endpoint'>
+
+export const offlineModeAuthStatus = {
+    endpoint: '',
+    isDotCom: true,
+    isLoggedIn: true,
+    isOfflineMode: true,
+    isFireworksTracingEnabled: false,
+    showInvalidAccessTokenError: false,
+    authenticated: true,
+    hasVerifiedEmail: true,
+    requiresVerifiedEmail: true,
+    siteHasCodyEnabled: true,
+    siteVersion: '',
+    userCanUpgrade: false,
+    username: 'offline',
+    primaryEmail: '',
+    displayName: '',
+    avatarURL: '',
+    codyApiVersion: 0,
+} satisfies AuthStatus