+++ vscode/test/fixtures/workspace/type.ts
@@ -1,0 +1,4 @@
+interface Fruit {
+    appleName: string
+    appleAge: number
+}